fix(sampling): accumulate top-p mass in float32
apply_top_p summed the sorted probabilities in the logprobs dtype. For
bfloat16 the running sum stops growing once each new term is below its
rounding step, so the tail never reaches 1 - top_p: on a 151,936-token
row, top_p=0.9 kept 312 tokens in bfloat16 against 35,960 in float32.

Compute the probabilities and their prefix sum in float32. The mask is
still applied to the caller's logprobs, so the output dtype is
unchanged. apply_top_p_top_k already works in float32.

+++ omlx/utils/sampling.py
@@ -25,7 +25,9 @@ import mlx.core as mx
 def apply_top_p(logprobs: mx.array, top_p: float) -> mx.array:
     """Top-p (nucleus) filtering — keep the smallest set of tokens whose
     cumulative probability mass is at least ``top_p``."""
-    probs = mx.exp(logprobs)
+    # Sum in float32: a bfloat16 running sum over a large vocabulary stops
+    # growing once each new term is below its rounding step.
+    probs = mx.exp(logprobs.astype(mx.float32))
     sorted_indices = mx.argsort(logprobs, axis=-1)
     sorted_probs = mx.take_along_axis(probs, sorted_indices, axis=-1)
